Class by different files
