+++ modules/sirotkin-nikita-finance_notepad/test/test_finance_notepad.cpp
@@ -6,273 +6,6 @@
 
 using std::string;
 
-TEST(Sirotkin_Nikita_FinanceNotepadTest, unfit_number_dates_is_set_to_default) {
-  // Arrange
-  string number_great = date(32, months::JANUARY, 2001).toString();
-  string expect = DEFAULT_DATE.toString();
-
-  // Act
-
-  // Assert
-  EXPECT_EQ(expect.compare(number_great), 0);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, great_year_dates_is_set_to_default) {
-  // Arrange
-  string year_great = date(1, months::JANUARY, 2101).toString();
-  string expect = DEFAULT_DATE.toString();
-
-  // Act
-
-  // Assert
-  EXPECT_EQ(expect.compare(year_great), 0);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, less_year_dates_is_set_to_default) {
-  // Arrange
-  string year_less = date(1, months::JANUARY, 1999).toString();
-  string expect = DEFAULT_DATE.toString();
-
-  // Act
-
-  // Assert
-  EXPECT_EQ(expect.compare(year_less), 0);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest,
-    dates_with_dif_numbers_can_compare_true) {
-  // Arrange
-  date number_less = date(1, months::JANUARY, 2001);
-  date number_great = date(3, months::JANUARY, 2001);
-
-  // Act
-
-  // Assert
-  EXPECT_TRUE(number_less < number_great);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest,
-    dates_with_dif_numbers_can_compare_false) {
-  // Arrange
-  date number_less = date(1, months::JANUARY, 2001);
-  date number_great = date(3, months::JANUARY, 2001);
-
-  // Act
-
-  // Assert
-  EXPECT_FALSE(number_less > number_great);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest,
-    dates_with_dif_months_can_compare_true) {
-  // Arrange
-  date month_less = date(1, months::JANUARY, 2001);
-  date month_great = date(1, months::MARCH, 2001);
-
-  // Act
-
-  // Assert
-  EXPECT_TRUE(month_great > month_less);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest,
-    dates_with_dif_months_can_compare_false) {
-  // Arrange
-  date month_less = date(1, months::JANUARY, 2001);
-  date month_great = date(1, months::MARCH, 2001);
-
-  // Act
-
-  // Assert
-  EXPECT_FALSE(month_great < month_less);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest,
-    dates_with_dif_year_can_compare_true) {
-  // Arrange
-  date year_less = date(1, months::JANUARY, 2001);
-  date year_great = date(1, months::JANUARY, 2002);
-
-  // Act
-
-  // Assert
-  EXPECT_TRUE(year_great > year_less);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest,
-    dates_with_dif_year_can_compare_false) {
-  // Arrange
-  date year_less = date(1, months::JANUARY, 2001);
-  date year_great = date(1, months::JANUARY, 2002);
-
-  // Act
-
-  // Assert
-  EXPECT_FALSE(year_great < year_less);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, dates__can_compare_by_eq_true) {
-  // Arrange
-  date first = MIN_DATE;
-  date second = MIN_DATE;
-
-  // Act
-
-  // Assert
-  EXPECT_TRUE(first == second);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, dates__can_compare_by_eq_false) {
-  // Arrange
-  date first = MIN_DATE;
-  date second = MAX_DATE;
-
-  // Act
-
-  // Assert
-  EXPECT_FALSE(first == second);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, table_can_get_category_id) {
-  // Arrange
-  category_table table;
-  std::string input("Food");
-  int expect = 5;
-  int result;
-
-  // Act
-  result = table.getIdOf(input);
-
-  // Assert
-  EXPECT_EQ(result, expect);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, table_can_get_category_name) {
-  // Arrange
-  category_table table;
-  int input = 5;
-  string expect = "Food";
-  string result;
-
-  // Act
-  result = table.getNameOf(input);
-
-  // Assert
-  EXPECT_EQ(result.compare(expect), 0);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest,
-    table_can_get_default_category_by_wrong_id) {
-  // Arrange
-  category_table table;
-  int input = table_size+1;
-  string expect = "Other";
-  string result;
-
-  // Act
-  result = table.getNameOf(input);
-
-  // Assert
-  EXPECT_EQ(result.compare(expect), 0);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest,
-    table_can_get_default_category_by_wrong_name) {
-  // Arrange
-  category_table table;
-  string input = "Something";
-  int expect = table_size-1;
-  int result;
-
-  // Act
-  result = table.getIdOf(input);
-
-  // Assert
-  EXPECT_EQ(result, expect);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, note_can_get_category_id) {
-  // Arrange
-  category_table table;
-  note debt(DEFAULT_DATE, 10.f, table.getIdOf("Debt"), &table, "Something");
-  int expect = table.getIdOf("Debt");
-  int result;
-
-  // Act
-  result = debt.getCategoriesId();
-
-  // Assert
-  EXPECT_EQ(result, expect);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, note_can_get_category_name) {
-  // Arrange
-  category_table table;
-  note debt(DEFAULT_DATE, 10.f, table.getIdOf("Debt"), &table, "Something");
-  string expect = "Debt";
-  string result;
-
-  // Act
-  result = debt.getCategory();
-
-  // Assert
-  EXPECT_EQ(result.compare(expect), 0);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, note_can_get_sum) {
-  // Arrange
-  category_table table;
-  note debt(DEFAULT_DATE, 10.f, table.getIdOf("Debt"), &table, "Something");
-  float expect = 10.f;
-  float result;
-
-  // Act
-  result = debt.getSum();
-
-  // Assert
-  EXPECT_FLOAT_EQ(result, expect);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, note_can_get_date) {
-  // Arrange
-  category_table table;
-  note debt(MAX_DATE, 10.f, table.getIdOf("Debt"), &table, "Something");
-  date expect = MAX_DATE;
-  date result;
-
-  // Act
-  result = debt.getDate();
-
-  // Assert
-  EXPECT_TRUE(expect == result);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, note_can_get_comment) {
-  // Arrange
-  category_table table;
-  note debt(MAX_DATE, 10.f, table.getIdOf("Debt"), &table, "Something");
-  string expect = "Something";
-  string result;
-
-  // Act
-  result = debt.getComment();
-
-  // Assert
-  EXPECT_EQ(result.compare(expect), 0);
-}
-
-TEST(Sirotkin_Nikita_FinanceNotepadTest, note_can_compare_less) {
-  // Arrange
-  category_table table;
-  note max_note(MAX_DATE, 10.f, table.getIdOf("Debt"), &table, "Something");
-  note min_note(MAX_DATE, 10.f, table.getIdOf("Debt"), &table, "Something");
-
-  // Act
-
-  // Assert
-  EXPECT_TRUE(min_note < max_note);
-}
-
 TEST(Sirotkin_Nikita_FinanceNotepadTest, can_get_pouch) {
   // Arrange
   notepad notes(13.f);
